+++ src/main/java/fr/insee/genesis/infrastructure/document/rawdata/RawResponseDocument.java
@@ -3,6 +3,8 @@ package fr.insee.genesis.infrastructure.document.rawdata;
 import lombok.Builder;
 import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.index.CompoundIndex;
+import org.springframework.data.mongodb.core.index.CompoundIndexes;
 import org.springframework.data.mongodb.core.index.IndexDirection;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
@@ -12,6 +14,9 @@ import java.util.Map;
 
 @Builder
 @Document(collection = "rawResponses")
+@CompoundIndexes({
+        @CompoundIndex(name = "payload_campaignId_index", def = "{ 'payload.campaignId': 1 }")
+})
 public record RawResponseDocument (
         @Id
         ObjectId id,
@@ -22,7 +27,5 @@ public record RawResponseDocument (
         Map<String,Object> payload,
         LocalDateTime recordDate,
         @Indexed(direction = IndexDirection.DESCENDING)
-        LocalDateTime processDate,
-        @Indexed
-        String campaignId
+        LocalDateTime processDate
 ){}